Renamed SquadList to SquadMenu
