+++ src/components/SquadMenu.vue
